backend/requirements.txt
# Core dependencies
selenium>=4.20.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0
requests>=2.31.0
requests[socks]>=2.31.0
Pillow>=10.0.0

# Audio processing (for reCAPTCHA audio challenges)
SpeechRecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.11

# Jupyter environment
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Database (ADICIONADO)
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Web scraping & parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Image processing
opencv-python-headless>=4.8.0
scikit-image>=0.21.0

# HTTP client improvements
httpx>=0.24.0
aiohttp>=3.8.0

# System utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Development utilities
rich>=13.0.0

# Music & Audio APIs
openai>=1.0.0
musicbrainzngs>=0.7.0
discogs-client>=2.3.0

# Visualization
matplotlib>=3.6.0

# Search APIs
google-search-results>=2.4.0

# Optional: 2CAPTCHA service (fallback)
2captcha-python>=1.1.0

pillow-heif
# Django & Django Ninja
django>=4.2.0
django-ninja>=1.0.0
django-cors-headers>=4.0.0
